requirements.txt

chromadb
transformers==4.57.6
speechbrain

webrtcvad
pydub

# Audio processing
soundfile
audioread
librosa
ffmpeg-python

# HTTP + serving
fastapi
uvicorn[standard]
requests
fastmcp

# Utils
numpy
scipy
tqdm
pyyaml
python-dotenv